+++ samples/Axent.ExampleApi/WelcomeRequest.cs
@@ -10,7 +10,7 @@ internal sealed record WelcomeRequest : IRequest<string>;
 
 internal sealed class WelcomeRequestHandler : IRequestHandler<WelcomeRequest, string>
 {
-    public ValueTask<Response<string>> HandleAsync(RequestContext<WelcomeRequest> context,
+    public ValueTask<Response<string>> HandleAsync(WelcomeRequest request,
         CancellationToken cancellationToken = default) =>
         ValueTask.FromResult(Response.Success("Axent Example Api is up and running!"));
 }