+++ src/Axent.Extensions.Authorization/AuthorizationPipe.cs
@@ -24,31 +24,31 @@ internal sealed class AuthorizationPipe<TRequest, TResponse>
     }
 
     public async ValueTask<Response<TResponse>> ProcessAsync(
-        IPipelineChain<TRequest, TResponse> chain,
-        RequestContext<TRequest> context,
+        TRequest request,
+        AxentPipelineContinuation<TRequest, TResponse> next,
         CancellationToken cancellationToken = default)
     {
         if (_metadata.AllowAnonymous || _metadata.AuthorizeData.Length == 0)
         {
-            return await chain.NextAsync(context, cancellationToken);
+            return await next(request, cancellationToken);
         }
 
         var policy = await AuthorizationPolicy.CombineAsync(_policyProvider, _metadata.AuthorizeData);
 
         if (policy is null)
         {
-            return await chain.NextAsync(context, cancellationToken);
+            return await next(request, cancellationToken);
         }
 
         var user = _principalAccessor.Principal;
         var authorizationResult = await _authorizationService.AuthorizeAsync(
             user,
-            context.Request,
+            request,
             policy);
 
         if (authorizationResult.Succeeded)
         {
-            return await chain.NextAsync(context, cancellationToken);
+            return await next(request, cancellationToken);
         }
 
         return Response.Failure(user.Identity?.IsAuthenticated == true