+++ src/Axent.Templates.MinimalApi/Program.cs
@@ -13,7 +13,7 @@ var app = builder.Build();
 
 app.MapGet("/", () => "API is up and running");
 
-app.MapGet("/api/weather-forecast", async (ISender sender, CancellationToken cancellationToken) =>
+app.MapGet("/api/weather-forecast", async (IRequestSender<WeatherForecastQuery, WeatherForecast[]> sender, CancellationToken cancellationToken) =>
 {
     var response = await sender.SendAsync(new WeatherForecastQuery(), cancellationToken);
     return response.ToResult();